requirements.txt
requests==2.32.3
urllib3==2.2.3
pydantic==2.9.2
PyYAML==6.0.2
loguru==0.7.2
tqdm==4.66.5
tenacity==9.0.0
pandas==2.2.3
openpyxl==3.1.5

# token-auth (альтернатива mTLS)
gigachat>=0.1.28
python-dotenv>=1.0.0

# dev / test
pytest==8.3.3
